pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "restormer-modular"
version = "1.5.0"
description = "A modular Restormer refactor with multi-model support, enhanced logging, Perceptual Metrics (LPIPS) & beautiful CLI evaluation tables, interactive visualization dashboard, and modernized uv-based dependency management with CUDA support."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Himesh" }
]
dependencies = [
    "torch==2.5.1",
    "torchvision==0.20.1",
    "torchaudio==2.5.1",
    "numpy==2.0.1",
    "opencv-python==4.12.0.88",
    "einops==0.8.1",
    "pyyaml==6.0.3",
    "tqdm==4.67.1",
    "scikit-image==0.26.0",
    "tensorboard==2.20.0",
    "timm==1.0.22",
    "PyWavelets==1.9.0",
    "fvcore==0.1.5.post20221221",
    "lpips==0.1.4",
    "gdown==5.2.0",
    "addict==2.4.0",
    "scipy==1.16.3",
    "matplotlib==3.10.8",
    "natsort==8.4.0",
    "h5py==3.15.1",
    "yacs==0.1.8",
    "yapf==0.43.0",
    "requests==2.32.5",
    "ninja==1.13.0",
    "numba==0.63.1",
    "scikit-learn==1.8.0",
    "future==1.0.0",
    "pillow==11.1.0",
    "gradio>=3.55",
    "tabulate>=0.9.0",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "core*", "data*"]
